Separated runRecording package
