+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         android:theme="@style/Theme.RunTracker"
         tools:targetApi="31">
         <activity
-            android:name=".MapActivity"
+            android:name=".runRecording.MapActivity"
             android:exported="false" />
         <activity
             android:name=".settings.SettingsActivity"
@@ -64,8 +64,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <service android:name=".TimerService"/>
-        <service android:name=".TrackerService"/>
+        <service android:name=".runRecording.TimerService"/>
+        <service android:name=".runRecording.TrackerService"/>
     </application>
 
 </manifest>
