+++ app/src/main/java/com/example/runtracker/runRecording/MapActivity.kt
@@ -1,7 +1,8 @@
-package com.example.runtracker
+package com.example.runtracker.runRecording
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.runtracker.R
 
 class MapActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
