+++ app/src/main/res/layout/fragment_map.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".MapFragment">
+    tools:context=".runRecording.MapFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/recyclerView"
